requirements.txt
# Python Version
# Python >= 3.8

# Core Data Science
pandas==2.2.3
numpy==2.3.2
scipy==1.16.1

# Statistical Modeling
# Pinned to 0.14.6 for reproducible regression analysis. 0.14.0 introduced breaking changes (deprecated GLM link names). See DEPENDENCIES.md for upgrade strategy.
statsmodels==0.14.6
linearmodels>=0.6.0
scikit-learn==1.7.2
lifelines==0.30.0

# Data Formats
PyYAML==6.0.2
# Pinned to 21.0.0 for Python 3.8-3.13 compatibility. 23.0.0+ requires Python >= 3.10. See DEPENDENCIES.md for upgrade strategy and performance notes.
pyarrow==21.0.0
openpyxl==3.1.5

# Utilities
psutil==7.2.1
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
python-dateutil==2.9.0.post0
# Optional: RapidFuzz for fuzzy string matching (Tier 3 entity linking)
# Install for improved entity match rates: pip install rapidfuzz>=3.14.0
# Pipeline runs without it (graceful degradation), but match rate is lower
# See DEPENDENCIES.md for details on optional dependencies and performance impact
rapidfuzz>=3.14.0

# Type stubs for mypy strict checking
pandas-stubs>=2.2.0  # Type stubs for pandas (modern replacement for types-pandas)
types-psutil>=6.0.0  # Type stubs for psutil
types-requests>=2.31.0  # Type stubs for requests (if used)
types-PyYAML>=6.0.0  # Type stubs for PyYAML

# Testing
pytest>=8.0.0
pytest-benchmark>=4.0.0
pytest-cov>=4.1.0
pytest-mypy>=0.10.0
